refactor(api): replace .tsx with .ts for weekly summary route
 The weekly summary route implementation has been refactored from a TypeScript JSX file (.tsx) to a standard
 TypeScript file (.ts). This change streamlines the codebase as JSX syntax is not required for the API route.
 All functionality remains intact with necessary adjustments to accommodate the TypeScript syntax.
